requirements.txt
matplotlib
numpy
biopython
xlrd==1.2.0
keras
tensorflow
